integer -> double for file size

+++ init-scripts/init-db.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS files(
     id serial primary key,
     url varchar(50),
-    size integer,
+    size double,
     content_type varchar(50),
     content bytea
 );
@@ -9,6 +9,6 @@ CREATE TABLE IF NOT EXISTS files(
 CREATE TABLE IF NOT EXISTS summary(
     id serial primary key,
     files_count integer,
-    files_size integer,
+    files_size double,
     updated_at timestamp
 );
